+++ docs/changelog.md
@@ -12,6 +12,9 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-pwa**
+  * Adds `WebAppHideToolbarFeature.toolbarVisible` to check the current toolbar visibility set by the feature.
+
 # 44.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v43.0.0...v44.0.0)
@@ -35,7 +38,7 @@ permalink: /changelog/
 * **feature-push**
   * Adds the `getSubscription` call to check if a subscription exists.
 
-* **browser-engine-gecko-***
+* **browser-engine-gecko-**
   * Fixes GeckoWebPushDelegate to gracefully return when a subscription is not available.
 
 * **feature-session**